requirements.txt
scikit-learn
scipy
numpy
pandas
sib-clustering
sentence_transformers
nltk
streamlit
requests
transformers
datasets
